+++ octave_transposer.hpp
@@ -2,8 +2,6 @@
 
 #include "midi_event.hpp"
 
-#include <optional>
-
 namespace zeta {
 
 class OctaveTransposer final {
@@ -11,11 +9,15 @@ public:
     void octaveDown() noexcept;
     void octaveUp() noexcept;
 
-    std::optional<MidiMessage> transpose(
-        const MidiMessage& message
-    ) const noexcept;
+    MidiMessage transpose(const MidiMessage& message) const noexcept;
 
 private:
+    static bool hasKey(MidiMessageType type) noexcept {
+        return type == MidiMessageType::NoteOff
+            || type == MidiMessageType::NoteOn
+            || type == MidiMessageType::PolyphonicKeyPressure;
+    }
+
     int octaves_{};
 };